+++ src/index.js
@@ -1,4 +1,4 @@
 export { TrackedProvider, createCustomContext } from './TrackedProvider';
-export { useTracked } from './useTracked';
+export { useTrackedState, useTracked } from './useTracked';
 export { useDispatch } from './useDispatch';
 export { useSelector } from './useSelector';